refactor: 색상값, 패딩값 등 :root에 상수화하기(app.css에) && 치료내역 모아보기 기능의 컨트롤러, 뷰, css파일 등 파일명들을 일관성 있게 통일(전부 treatment로 통일)
