+++ src/main/webapp/WEB-INF/views/common/header.jsp
@@ -20,7 +20,8 @@
 		href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css"
     />
     <link rel="stylesheet" href="/springframework-mini-project/resources/css/reset.css" />
-    <link rel="stylesheet" href="/springframework-mini-project/resources/css/medical-history-page.css" />
+    <link rel="stylesheet" href="/springframework-mini-project/resources/css/app.css" />
+    <link rel="stylesheet" href="/springframework-mini-project/resources/css/treatment.css" />
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/js/bootstrap.bundle.min.js"></script>
     <style>
@@ -294,7 +295,7 @@ margin-left:0.5rem;
 			            <div id="ham-menu-wrap">
 			                <div class="ham-menu mt-5">
 			                    <ul class="ham-top-menu">
-			                        <li><a href="${pageContext.request.contextPath}/treatmentlist/main"><i class="fa-solid fa-tooth"></i> 치료 내역 모아보기<i class="bi bi-caret-down-fill"></i></a>
+			                        <li><a href="${pageContext.request.contextPath}/treatment/main"><i class="fa-solid fa-tooth"></i> 치료 내역 모아보기<i class="bi bi-caret-down-fill"></i></a>
 
 			                        </li>
 			                    </ul>
