+++ src/main/webapp/resources/css/reset.css
@@ -139,6 +139,3 @@ a:active {
   color: inherit;
   text-decoration: none;
 }
-body{
-	letter-spacing: -0.05rem;
-}
