+++ src/main/webapp/resources/css/treatment.css
@@ -107,57 +107,3 @@
 	width: 100%;
 	margin-bottom: 0.3rem;
 }
-
-.btn-large-solid {
-	border: 1.5px solid var(--osstem-orange);
-	background-color: var(--osstem-orange);
-	border-radius: 3px;
-	padding: 1rem;
-	text-align: center;
-	font-weight: 600;
-	color: white;
-	margin-bottom: 0.8rem;
-}
-.btn-large-solid:hover {
-	color: white;
-	background-color: var(--osstem-orange-dark);
-}
-.btn-large-hollow {
-	border: 1.5px solid var(--osstem-orange);
-	background-color: var(--pale-peach);
-	border-radius: 3px;
-	padding: 1rem;
-	text-align: center;
-	font-weight: 600;
-	color: var(--osstem-orange);
-	margin-bottom: 0.8rem;
-}
-.btn-large-hollow:hover {
-	color: var(--osstem-orange);
-	background-color: var(--pale-peach-dark);
-}
-
-:root {
-	/* signiture colors */
-	--osstem-orange: rgb(242, 101, 34);
-	--osstem-orange-dark: rgb(226, 93, 32);
-	--osstem-gray: rgb(147, 147, 138);
-	--pale-peach: #FBF7F2;
-	--pale-peach-dark: #f0ebe6;
-
-	/* font colors */
-	--almost-black-gray: rgb(37, 40, 43);
-	--dark-gray: rgb(84, 91, 98);
-	--gray: rgb(108, 117, 125);
-	--light-gray: rgb(166, 166, 166);
-	--ligther-gray: rgb(191, 191, 191);
-
-	/* background colors */
-	--background-gray: rgb(250, 250, 250);
-	--border-gray: rgb(238, 238, 238);
-	--background-divider-gray: rgb(242, 242, 242);
-	--card-border: 2px solid var(--border-gray);
-
-	/* paddings */
-	--default-inner-space: 1rem;
-}
